requirements.txt
discord.py
python-dotenv
pandas
requests
ta
